admin: extract config dump handler to separate class (#12942)
