+++ source/server/admin/BUILD
@@ -15,6 +15,7 @@ envoy_cc_library(
     deps = [
         ":admin_filter_lib",
         ":clusters_handler_lib",
+        ":config_dump_handler_lib",
         ":config_tracker_lib",
         ":listeners_handler_lib",
         ":logs_handler_lib",
@@ -62,11 +63,6 @@ envoy_cc_library(
         "//source/common/router:scoped_config_lib",
         "//source/common/stats:isolated_store_lib",
         "//source/extensions/access_loggers/file:file_access_log_lib",
-        "@envoy_api//envoy/admin/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/endpoint/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/route/v3:pkg_cc_proto",
-        "@envoy_api//envoy/extensions/filters/network/http_connection_manager/v3:pkg_cc_proto",
     ],
 )
 
@@ -240,6 +236,26 @@ envoy_cc_library(
         "//source/common/http:header_map_lib",
         "//source/common/upstream:host_utility_lib",
         "@envoy_api//envoy/admin/v3:pkg_cc_proto",
+    ],
+)
+
+envoy_cc_library(
+    name = "config_dump_handler_lib",
+    srcs = ["config_dump_handler.cc"],
+    hdrs = ["config_dump_handler.h"],
+    deps = [
+        ":config_tracker_lib",
+        ":handler_ctx_lib",
+        ":utils_lib",
+        "//include/envoy/http:codes_interface",
+        "//include/envoy/server:admin_interface",
+        "//include/envoy/server:instance_interface",
+        "//source/common/buffer:buffer_lib",
+        "//source/common/http:codes_lib",
+        "//source/common/http:header_map_lib",
+        "@envoy_api//envoy/admin/v3:pkg_cc_proto",
+        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
+        "@envoy_api//envoy/config/endpoint/v3:pkg_cc_proto",
     ],
 )
